+++ fellow/commands/make_plan.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 
 from fellow.commands import CommandInput
-from fellow.commands.command import CommandContext
+from fellow.commands.Command import CommandContext
 
 
 class MakePlanInput(CommandInput):
@@ -12,13 +12,5 @@ def make_plan(args: MakePlanInput, context: CommandContext) -> str:
     """
     Creates a plan for the AI to follow. The plan will be in every future message for guidance.
     """
-    context["ai_client"].system_content.append(
-        {
-            "role": "system",
-            "content": args.plan,
-            "tokens": context["ai_client"].count_tokens(
-                {"role": "system", "content": args.plan}
-            ),
-        }
-    )
+    context["ai_client"].set_plan(args.plan)
     return "[OK] Plan created"