+++ tests/clients/test_OpenAIClient.py
@@ -4,14 +4,21 @@ from tempfile import NamedTemporaryFile
 from unittest.mock import MagicMock, patch
 
 import pytest
-from openai import NOT_GIVEN
 
-from fellow.clients.OpenAIClient import OpenAIClient
+from fellow.clients.OpenAIClient import OpenAIClient, OpenAIClientConfig
+from fellow.commands import Command, ViewFileInput, view_file
 
 
 @pytest.fixture
 def client():
-    return OpenAIClient(system_content="You are a helpful assistant.")
+    return OpenAIClient.create(
+        OpenAIClientConfig(
+            system_content="You are a helpful assistant.",
+            memory_max_tokens=1000,
+            summary_memory_max_tokens=1000,
+            model="gpt-3.5-turbo",
+        )
+    )
 
 
 @pytest.fixture
@@ -80,8 +87,14 @@ def test_chat(mock_create, client, mock_openai_api_key):
         {"name": "func", "description": "func1 description", "parameters": {"arg": 1}}
     ]
 
-    response = client.chat("Hi there", NOT_GIVEN, functions=functions)
-    assert response == ("Hello!", None, None)
+    response = client.chat(
+        functions=functions, message="Hi there", function_result=None
+    )
+    assert response == {
+        "message": "Hello!",
+        "function_name": None,
+        "function_args": None,
+    }
     assert len(client.memory) == 2
     assert client.memory[-1]["content"] == "Hello!"
     assert "tokens" in client.memory[-1]
@@ -96,7 +109,11 @@ def test_chat(mock_create, client, mock_openai_api_key):
     response = client.chat(
         "", function_result={"name": "get_code", "output": "import os"}
     )
-    assert response == (None, "get_code", "{}")
+    assert response == {
+        "message": None,
+        "function_name": "get_code",
+        "function_args": "{}",
+    }
 
 
 @patch.object(OpenAIClient, "_summarize_memory")
@@ -108,7 +125,7 @@ def test_memory_summarization_triggered(mock_summarize, client):
         {"role": "user", "content": f"msg{i}", "tokens": 300} for i in range(5)
     ]
     client.summary_memory = []
-    client.count_tokens = lambda _: 300
+    client._count_tokens = lambda _: 300
     client.memory_max_tokens = 1000
     client.summary_memory_max_tokens = 1000
 
@@ -220,7 +237,7 @@ def test__maybe_summarize_memory(client):
     ] * 3  # 180 tokens total
 
     # Patch count_tokens to return fixed value for simplicity
-    client.count_tokens = lambda msg: 10
+    client._count_tokens = lambda msg: 10
 
     # Trigger summarization
     client._maybe_summarize_memory()
@@ -248,3 +265,48 @@ def test__maybe_summarize_memory(client):
         in client.summary_memory[1]["content"]
     )
     assert len(mock_summarize.call_args[0][0]) == 2
+
+
+def test_set_plan():
+    # todo: ...
+    pass
+
+
+def test_command_get_function_schema(client):
+    command = Command(ViewFileInput, view_file)
+    assert client.get_function_schema(command) == {
+        "name": "view_file",
+        "description": "View the contents of a file, optionally between specific line numbers.",
+        "parameters": {
+            "properties": {
+                "filepath": {
+                    "description": "The path to the file to be viewed.",
+                    "title": "Filepath",
+                    "type": "string",
+                },
+                "from_line": {
+                    "anyOf": [{"type": "integer"}, {"type": "null"}],
+                    "default": None,
+                    "description": "Optional 1-based starting line number.",
+                    "title": "From Line",
+                },
+                "to_line": {
+                    "anyOf": [{"type": "integer"}, {"type": "null"}],
+                    "default": None,
+                    "description": "Optional 1-based ending line number (inclusive).",
+                    "title": "To Line",
+                },
+            },
+            "required": ["filepath"],
+            "title": "ViewFileInput",
+            "type": "object",
+        },
+    }
+
+    with pytest.raises(ValueError) as err:
+        client.get_function_schema(Command(ViewFileInput, lambda x, y: None))
+    assert str(err.value) == "[ERROR] Command handler is __doc__ is empty"
+
+    with pytest.raises(ValueError) as err:
+        client.get_function_schema(Command(ViewFileInput, "no-name"))
+    assert str(err.value) == "[ERROR] Command handler is not callable with __name__."